feat(ci): CodeRabbit no lugar do Greptile como revisor de PR
Gate, labels e automerge trocados de needs-greptile-resolution/greptile-resolved
para needs-coderabbit-resolution/coderabbit-resolved. O gate agora casa por
check e autor coderabbitai — não mais pelo corpo do comentário, para não
bloquear PR quando humano menciona o revisor. .coderabbit.yaml em pt-BR,
perfil chill, sem revisar draft. Instalação do GitHub App fica fora do git.

Agent: opencode (GLM-5.2)
