+++ .coderabbit.yaml
@@ -1,0 +1,14 @@
+# Substitui o Greptile como revisor de PR (mantido pelo portão needs-coderabbit-resolution
+# no workflow csbrasil-bot-pr-classify). Docs: https://docs.coderabbit.io/reference/configuration
+language: pt-BR
+reviews:
+  profile: chill
+  request_changes_workflow: false
+  high_level_summary: true
+  review_status: true
+  poem: false
+  auto_review:
+    enabled: true
+    drafts: false
+chat:
+  auto_reply: true
