+++ scripts/ci/check_automerge.py
@@ -22,19 +22,19 @@ def main() -> int:
     labels = {l["name"] for l in pr.get("labels", [])}
     files = [f.get("path", "") for f in pr.get("files", [])]
     touches_workflows = any(path.startswith(".github/workflows/") for path in files)
-    greptile_blocked = "needs-greptile-resolution" in labels and "greptile-resolved" not in labels
+    coderabbit_blocked = "needs-coderabbit-resolution" in labels and "coderabbit-resolved" not in labels
     eligible = (
         not pr.get("isDraft", False)
         and "safe-automerge" in labels
         and not touches_workflows
-        and not greptile_blocked
+        and not coderabbit_blocked
         and pr.get("reviewDecision") != "CHANGES_REQUESTED"
         and pr.get("mergeStateStatus") in {"CLEAN", "HAS_HOOKS"}
         and check_rollup_ok(pr.get("statusCheckRollup", []))
     )
     print(json.dumps({
         "eligible": eligible,
-        "greptile_blocked": greptile_blocked,
+        "coderabbit_blocked": coderabbit_blocked,
         "touches_workflows": touches_workflows,
     }))
     return 0
