+++ scripts/ci/ensure_labels.py
@@ -10,8 +10,8 @@ LABELS = {
     "target:main": ("0366d6", "PR apontando para a branch main"),
     "target:staging": ("0e8a16", "PR apontando para a branch staging"),
     "target:release": ("5319e7", "PR apontando para uma branch de release"),
-    "needs-greptile-resolution": ("b60205", "PR não pode ser mergeada enquanto houver pendência do Greptile"),
-    "greptile-resolved": ("0e8a16", "Maintainer confirmou que os apontamentos do Greptile foram resolvidos"),
+    "needs-coderabbit-resolution": ("b60205", "PR não pode ser mergeada enquanto houver pendência do CodeRabbit"),
+    "coderabbit-resolved": ("0e8a16", "Maintainer confirmou que os apontamentos do CodeRabbit foram resolvidos"),
     "bot-fixable": ("0e8a16", "Issue ou PR elegível para tentativa automatizada de correção"),
     "covered-by-pr": ("fbca04", "Issue já atacada por PR aberta"),
     "crash-auto": ("B60205", "crash de produção reportado por /api/jserror ou prod-watch"),
